backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
openai==1.30.1
anthropic==0.28.0
google-api-python-client==2.131.0
google-auth==2.29.0
loguru==0.7.2
